chore: Update Action versions, fix some gosec things
Squashed commit of the following:

commit 5ce03fec14aeeef784ee07da5298eac8eb698062
Author: Rafael Espinoza <rafael@rafaelespinoza.com>
Date:   Sun Sep 29 19:26:38 2024 -0700

    fix: Adjust typing to address some gosec violations

    It's been several months since CI last ran. In that time, some upstream
    gosec rules have changed. These were caught in CI run #23
    at https://github.com/rafaelespinoza/notexfr/actions/runs/11098075302/job/30830225357.

    Breakage:
        [/github/workspace/internal/repo/edam/notes.go:43] - G115 (CWE-190): integer overflow conversion int -> int32 (Confidence: MEDIUM, Severity: HIGH)
            42: 	pageSize := int32(n.rqp.PageSize)
          > 43: 	pagination := newPaginator(int32(n.rqp.LoIndex), int32(n.rqp.HiIndex))
            44:

        Autofix:

        [/github/workspace/internal/repo/edam/notes.go:43] - G115 (CWE-190): integer overflow conversion int -> int32 (Confidence: MEDIUM, Severity: HIGH)
            42: 	pageSize := int32(n.rqp.PageSize)
          > 43: 	pagination := newPaginator(int32(n.rqp.LoIndex), int32(n.rqp.HiIndex))
            44:

        Autofix:

        [/github/workspace/internal/repo/edam/notes.go:42] - G115 (CWE-190): integer overflow conversion int -> int32 (Confidence: MEDIUM, Severity: HIGH)
            41: 	filter := n.rqp.toFilter()
          > 42: 	pageSize := int32(n.rqp.PageSize)
            43: 	pagination := newPaginator(int32(n.rqp.LoIndex), int32(n.rqp.HiIndex))

    Addressed this by changing the types of fields on the NotesRemoteQueryParams
    struct to be the same types expected by the library github.com/dreampuf/evernote-sdk-golang.
    The issue is just shifted over to the cmd package now; oddly gosec didn't catch that.
    This could be validated early. Added a TODO

commit b9357789d5d04770167cf07f7e594c9252a52596
Author: Rafael Espinoza <rafael@rafaelespinoza.com>
Date:   Sun Sep 29 18:55:49 2024 -0700

    ci: Update action versions

+++ internal/repo/edam/notes.go
@@ -39,8 +39,8 @@ func (n *Notes) FetchRemote(ctx context.Context) (out []entity.LinkID, err error
 		return
 	}
 	filter := n.rqp.toFilter()
-	pageSize := int32(n.rqp.PageSize)
-	pagination := newPaginator(int32(n.rqp.LoIndex), int32(n.rqp.HiIndex))
+	pageSize := n.rqp.PageSize
+	pagination := newPaginator(n.rqp.LoIndex, n.rqp.HiIndex)
 
 	yes := true
 	// resultSpec tells evernote which fields to include in the search. By
@@ -126,9 +126,9 @@ func (n *Notes) FetchRemote(ctx context.Context) (out []entity.LinkID, err error
 
 // NotesRemoteQueryParams is a set of named options for listing Evernote notes.
 type NotesRemoteQueryParams struct {
-	LoIndex    int
-	HiIndex    int
-	PageSize   int
+	LoIndex    int32
+	HiIndex    int32
+	PageSize   int32
 	TagIDs     []string
 	NotebookID string
 	Verbose    bool
